I think I found the problem
The theme list was inheriting its "align-items" property from its parent, which had everything centered. Also, I simplified the theme list, so that it's a ul now instead of a bunch of divs. Furthermore, I re-added some padding on the left side of the theme selector button that got lost when I switched formats.